requirements.txt
aiohttp
pyfiglet
psutil
tensorboard
bitsandbytes
keyboard
InquirerPy
prompt_toolkit
argparse
asyncio
datasets
packaging
python-dotenv
trl
prettytable
loguru
GPUtil
pandas
requests
scikit-learn
tqdm
huggingface_hub
pydantic
rich
safetensors
wandb
sentencepiece
jinja2
accelerate
bitsandbytes
sentence-transformers
tokenizers
peft
keylimiter
pydantic-settings
evaluate
torch
nltk
matplotlib
seaborn
scikit-learn
wandb
communex
firebase_admin

